requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy==1.19.2
pandas
pandas-profiling==3.0.0
scikit-learn==0.22

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs==2021.7.0
google-cloud-storage==1.42.0
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor
